Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
indic-transliteration==2.3.64
transformers==4.47.1
torch==2.5.1
nltk==3.9.1
sentencepiece==0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
